update the mixed_precision_dtype

+++ train.py
@@ -314,7 +314,7 @@ if __name__ == "__main__":
     num_epochs = 10
     batch_size = 4
     gradient_accumulation_steps = 16
-    mixed_precision_dtype = torch.bfloat16
+    mixed_precision_dtype = torch.float16
 
     learning_rate = 3e-4
     lr_scheduler_type = 'cosine'
